updating mccormick api
